feat: add repository field to package.json
This commit introduces the addition of the repository field within the package.json file. This update specifies the type of version control (git in this case) and provides the URL to the repository where this package is hosted, https://github.com/JonDotsoy/kv.git. The inclusion of the repository information is aimed at enhancing project visibility and accessibility for users and contributors by directly linking the source code repository within the package metadata. This change facilitates easier navigation to the project's source code, issue tracking, and contributions.