feat(package.json): bump package version to v0.1.1 and add metadata
The version of `@jondotsoy/open-kv` has been updated from v0.1.0 to v0.1.1 to indicate the addition of new features or fix in a backwards-compatible manner. Alongside the version bump, the package.json file now includes a `description` about the package and specifies an `MIT` license. These metadata additions provide clearer insight into the purpose of the package and its open-source usage terms, helping users and contributors understand the project scope and legal considerations. These changes do not affect the functionality of the existing code but are intended to improve the package's usability and compliance.