chore: bump version to v0.1.2 in package.json
This commit updates the version of the @jondotsoy/open-kv package from v0.1.1 to v0.1.2. The bump in version reflects recent enhancements or fixes made to the package. This update ensures that users of the package will have access to the latest features and improvements. Updating the version number in the package.json file is a crucial step for package management and deployment processes, enabling a consistent and reliable usage experience for end-users.